+++ requirements.txt
@@ -1,3 +1,5 @@
 flask
 flask-sqlalchemy
 flask-admin
+pyyaml
+pyjwt
