+++ services/chatroom_service/config/chatroom_config.yaml
@@ -1,0 +1,7 @@
+database:
+  uri: "sqlite:///chatrooms.db"
+  track_modifications: false
+
+profile-service:
+  url: "localhost:5001"
+
