+++ services/profile_service/config/profile_config.yaml
@@ -1,0 +1,10 @@
+database:
+  uri: "sqlite:///profile.db"
+  track_modifications: false
+
+flask:
+  secret_key: 'ECE4150-Lab4'
+
+chatroom-service:
+  url: "localhost:5002"
+
